Refactor: move new arrivals and shopByBrand into Components folder
